pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-audit-kit"
version = "0.2.0"
description = "Security scanner for MCP-connected AI agent pipelines"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "AgentAuditKit Contributors"}
]
keywords = ["mcp", "security", "scanner", "ai-agents", "claude-code"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
agent-audit-kit = "agent_audit_kit.cli:cli"

[project.urls]
Homepage = "https://github.com/sattyamjjain/agent-audit-kit"
Repository = "https://github.com/sattyamjjain/agent-audit-kit"
Issues = "https://github.com/sattyamjjain/agent-audit-kit/issues"

[tool.setuptools.packages.find]
include = ["agent_audit_kit*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
